% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComboPlotFunc.R
\name{ComboPlotFunc}
\alias{ComboPlotFunc}
\title{Create the multi-plot figure for Water Quality Briefs}
\usage{
ComboPlotFunc(
  data1,
  data2,
  type = "Non-Marine",
  years = c(2009, 2011, 2013, 2015, 2017, 2019, 2021),
  yearrange = c(2008, 2022),
  graphtitle = "Water Quality at Selected Park and Resource",
  errorbartype = "CI"
)
}
\arguments{
\item{data1}{A dataset in the format of kahocentralnutdata. Must include the fields parameter, year, quarter, year_quarter, mean_value and (for marine only) depth.}

\item{data2}{A dataset in the format of kahocentralysidata. Must include the fields parameter, year, quarter, year_quarter, corrected_mean and (for marine only) depth.}

\item{type}{Resource type. Marine or Non-Marine.}

\item{years}{A list of years to be used as breaks on the graphs. Default is c(2009,2011,2013,2015,2017,2019,2021)}

\item{yearrange}{A list of length two indicating the min and max years displayed on the graph axes. Defaults is set to c(2008,2022).}

\item{graphtitle}{The desired graph title. Default is set to "Water Quality at Selected Park and Resource".}

\item{errorbartype}{The error bar type desired. Default is "CI" for 95\% confidence internval. Otherwise it will do Standard Error.}
}
\value{
A multi-plot of mean values for 9 parameters over time in the PACN desired format.
}
\description{
Creates a 9 plot figure with nutrient and ysi plots, suitable for publication
}
