% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdb_table_wq.R
\name{gdb_table_wq}
\alias{gdb_table_wq}
\title{Make a table from Geodatabase for photo processing}
\usage{
gdb_table_wq(gdb_name, gdb_location, gdb_layer)
}
\arguments{
\item{gdb_name}{the name of the geodatabase}

\item{gdb_location}{the folder the geodatabase is store in. Typically "geodatabase".}

\item{gdb_layer}{the name of the layer in the geodatabase. This appears to be the file description used for download.}
}
\value{
returns a table with all the observations in the geodatabase and 30 columns
}
\description{
Makes a table from the specified geodatabase in the format that is required by the watermark_wq function
}
