% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_watermark_wq.R
\name{process_watermark_wq}
\alias{process_watermark_wq}
\title{Process and Watermark WQ photos}
\usage{
process_watermark_wq(
  gdb_name,
  gdb_location,
  gdb_layer,
  park,
  loctype,
  select_months = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12),
  output_folder = "watermarked"
)
}
\arguments{
\item{gdb_name}{the name of the geodatabase}

\item{gdb_location}{the folder the geodatabase is store in. Typically "geodatabase".}

\item{gdb_layer}{the name of the layer in the geodatabase. This appears to be the file description used for download.}

\item{park}{the unit code desired}

\item{loctype}{location type for photos to be processed. AP, MR, BB, FW.}

\item{select_months}{a list of months you want processed}

\item{output_folder}{where you want the photos to go. Default is "watermarked". A new folder will be created if it does not exist.}
}
\value{
returns a table with all the observations in the geodatabase and 30 columns
}
\description{
Combines the gdb_table_wq and watermark_wq functions. Allowing for the user to select the park, location type, and month of sampling
}
