% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAvgNutPlotColorBy.R
\name{QAvgNutPlotColorBy}
\alias{QAvgNutPlotColorBy}
\title{Graph PACN nutrient data for non-marine resources colored by station or location}
\usage{
QAvgNutPlotColorBy(
  x,
  param,
  main.title,
  y.label,
  axis.digits = 1,
  colorby = "loc_name",
  errorbartype = "CI"
)
}
\arguments{
\item{x}{A dataset in the format of nutdata. Must include the fields parameter, year, quarter, year_quarter,mean_value.}

\item{param}{A parameter name. Typically TDN, TDP, NO3.}

\item{main.title}{The desired graph title.}

\item{y.label}{The desired y axis label}

\item{axis.digits}{The number of digits desired for the y axis.}

\item{colorby}{The variable you'd like to color by either station_id or loc_name}

\item{errorbartype}{The type of error bar desired. Default is "CI" for 95\% confidence interval. Otherwise it will do standard error.}
}
\value{
A plot of mean values over time in the PACN desired format.
}
\description{
Nutrient data are graphed for nutrient parameters
}
