% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAvgYsiPlot_marine.R
\name{QAvgYsiPlot_marine}
\alias{QAvgYsiPlot_marine}
\title{Graph PACN ysi data for marine resources}
\usage{
QAvgYsiPlot_marine(
  x,
  param,
  main.title,
  y.label,
  axis.digits,
  errorbartype = "CI"
)
}
\arguments{
\item{x}{A dataset in the format of ysidata. Must include the fields parameter, year, quarter, year_quarter, depth, corrected_mean.}

\item{param}{A parameter name. Typically Chl, pH, Sal, ODOConc, ODOSat, SpCond, Temp, and Turbid+.}

\item{main.title}{The desired graph title.}

\item{y.label}{The desired y axis label}

\item{axis.digits}{The number of digits desired for the y axis.}

\item{errorbartype}{The type of error bar desired. Default is "CI" for 95\% confidence interval. Otherwise it will do standard error.}
}
\value{
A plot of mean values over time in the PACN desired format.
}
\description{
YSI data are graphs for YSI parameters
}
