% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAvgYsiPlotColorBy.R
\name{QAvgYsiPlotColorBy}
\alias{QAvgYsiPlotColorBy}
\title{Graph PACN ysi data for non-marine resources colored by station or location}
\usage{
QAvgYsiPlotColorBy(
  x,
  param,
  main.title,
  y.label,
  axis.digits,
  colorby = "loc_name",
  errorbartype = "CI"
)
}
\arguments{
\item{x}{A dataset in the format of ysidata. Must include the fields parameter, year, quarter, year_quarter, depth, corrected_mean.}

\item{param}{A parameter name. Typically Chl, pH, Sal, ODOConc, ODOSat, SpCond, Temp, and Turbid+.}

\item{main.title}{The desired graph title.}

\item{y.label}{The desired y axis label}

\item{axis.digits}{The number of digits desired for the y axis.}

\item{colorby}{The variable you'd like to colorby either loc_name or station_id}

\item{errorbartype}{The type of error bar desired. Default is "CI" for 95\% confidence interval. Otherwise it will do standard error.}
}
\value{
A plot of mean values over time in the PACN desired format.
}
\description{
YSI data are graphed for YSI parameters
}
