% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Coordinates to Columns}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{x}{the sf table that needs to be separated}

\item{names}{the column names that will be used for the coordinates}
}
\value{
returns a table with two columns for x and y coordinates
}
\description{
Extracts coordinates from the sf object and put it in separate columns for X and Y
}
