% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark_streams.R
\name{watermark_streams}
\alias{watermark_streams}
\title{Watermark stream animals photos}
\usage{
watermark_streams(x, new_folder)
}
\arguments{
\item{x}{the table created from the geodatabase. Typically the output of gdb_table_wq().}

\item{new_folder}{the folder name you'd like the watermarked photos to go}
}
\value{
photos are processed and put in the new_folder, nothing is returned
}
\description{
Takes a table including photos in "DATA" and processes the photos.
It adds watermarks to the photos in the format required for PACN water quality.
}
