% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAvgNutPlot.R
\name{QAvgNutPlot}
\alias{QAvgNutPlot}
\title{Graph PACN nutrient data for non-marine resources}
\usage{
QAvgNutPlot(x, param, main.title, y.label, axis.digits = 1)
}
\arguments{
\item{x}{A dataset in the format of nutdata. Must include the fields parameter, year, quarter, year_quarter,mean_value.}

\item{param}{A parameter name. Typically TDN, TDP, NO3.}

\item{main.title}{The desired graph title.}

\item{y.label}{The desired y axis label}

\item{axis.digits}{The number of digits desired for the y axis.}
}
\value{
A plot of mean values over time in the PACN desired format.
}
\description{
Nutrient data are graphs for nutrient parameters
}
